Update readme to reflect order of output
The bullet list at the beginning of the README had the items listed in a different order than how they appear in the actual results. Also, in the "supported outputs" section, it wasn't clear that there is a default output type in the terminal (not CLI-table). I hope that makes sense.